Split bandwidth limiter tests into module
